Create handle request from digital specimen
