+++ src/test/java/eu/dissco/core/digitalspecimenprocessor/utils/TestUtils.java
@@ -1,6 +1,6 @@
 package eu.dissco.core.digitalspecimenprocessor.utils;
 
-import static eu.dissco.core.digitalspecimenprocessor.domain.FdoProfile.PRIMARY_SPECIMEN_OBJECT_ID;
+import static eu.dissco.core.digitalspecimenprocessor.domain.FdoProfileAttributes.PRIMARY_SPECIMEN_OBJECT_ID;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
